Added recent logs
