+++ static/js/src/demolist.tsx
@@ -1,184 +1,9 @@
 import React from "react";
 import ReactDOM from "react-dom";
+
+import DemoCard from "./demoCard";
 import DemoSearch from "./demoSearch";
-import { DemoUpdateStates, DemoService, DemoStates } from "./types";
-
-const CardStatus = (status: DemoStates) => {
-  if (status === DemoStates.RUNNING) {
-    return <div className="p-status-label--positive">Running</div>;
-  } else if (status === DemoStates.BUILDING) {
-    return <div className="p-status-label--caution">Builing</div>;
-  } else if (status === DemoStates.FAILED) {
-    return <div className="p-status-label--negative">Failed</div>;
-  }
-  return <div className="p-status-label">Unknown</div>;
-};
-
-const DemoButton = (props: {
-  name: string;
-  updateState: DemoUpdateStates;
-  onClick: (name: string, updateState: DemoUpdateStates) => Promise<void>;
-  disabled?: boolean;
-}) => {
-  const [isLoading, setIsLoading] = React.useState(false);
-
-  const handleClick = () => {
-    setIsLoading(true);
-
-    props.onClick(props.name, props.updateState).then(() => {
-      setIsLoading(false);
-    });
-  };
-
-  if (isLoading) {
-    return (
-      <button
-        disabled=""
-        class={`${
-          props.updateState == DemoUpdateStates.DELETE
-            ? "p-button--negative"
-            : "p-button--positive"
-        } is-processing`}
-      >
-        <i class="p-icon--spinner u-animation--spin is-light"></i>
-      </button>
-    );
-  }
-  return (
-    <button
-      className={`${
-        props.updateState == DemoUpdateStates.DELETE
-          ? "p-button--negative"
-          : "p-button--positive"
-      } js-processing-button`}
-      onClick={handleClick}
-      disabled={props.disabled}
-    >
-      <span id="button-label">
-        {props.updateState == DemoUpdateStates.DELETE
-          ? "Stop demo"
-          : "Restart demo"}
-      </span>
-    </button>
-  );
-};
-
-const DemoCard = (props: {
-  demo: DemoService;
-  handleSubmit: (name: string, state: DemoUpdateStates) => Promise<void>;
-}) => {
-  const { demo, handleSubmit } = props;
-
-  return (
-    <div className="p-card">
-      <div
-        style={{
-          display: "flex",
-          flexDirection: "row",
-          alignItems: "center",
-          justifyContent: "space-between",
-        }}
-      >
-        <div
-          style={{
-            display: "flex",
-            alignItems: "center",
-            justifyContent: "center",
-          }}
-        >
-          {CardStatus(demo.status)}
-        </div>
-        <div
-          style={{
-            display: "flex",
-            alignItems: "center",
-            justifyContent: "center",
-          }}
-        >
-          <p className="p-heading--4">{demo.running_time}</p>
-        </div>
-      </div>
-      <div
-        style={{
-          display: "flex",
-          flexDirection: "row",
-          alignItems: "center",
-          justifyContent: "space-between",
-        }}
-      >
-        <div
-          style={{
-            display: "flex",
-            alignItems: "center",
-            justifyContent: "center",
-          }}
-        >
-          <a>
-            <h3>{demo.name}</h3>
-          </a>
-        </div>
-      </div>
-
-      <div
-        style={{
-          display: "flex",
-          flexDirection: "row",
-          alignItems: "center",
-          justifyContent: "space-between",
-        }}
-      >
-        <div
-          style={{
-            display: "flex",
-            alignItems: "center",
-            justifyContent: "center",
-          }}
-        >
-          <div className="p-chip">
-            <a href={demo.pr_url}>
-              <img
-                height="32"
-                width="32"
-                src="static/images/gh.png"
-                alt={demo.name}
-              />
-            </a>
-          </div>
-          <div className="p-chip">
-            <a href={demo.host}>
-              <img
-                height="32"
-                width="32"
-                src="static/images/jenkins.png"
-                alt={demo.name}
-              />
-            </a>
-          </div>
-        </div>
-        <div
-          style={{
-            display: "flex",
-            alignItems: "center",
-            justifyContent: "center",
-          }}
-        >
-          <DemoButton
-            name={demo.name}
-            updateState={DemoUpdateStates.DELETE}
-            onClick={handleSubmit}
-            disabled={demo.status === DemoStates.FAILED}
-          />
-          <DemoButton
-            name={demo.name}
-            updateState={DemoUpdateStates.RESTART}
-            onClick={handleSubmit}
-            disabled={demo.status === DemoStates.RUNNING}
-          />
-        </div>
-      </div>
-    </div>
-  );
-};
+import { DemoUpdateStates, DemoService } from "./types";
 
 function App() {
   const [demos, setDemos] = React.useState([]);
